Partial cleanup.  Reduce timeouts and delays a bit.
