+++ api/token_refresh_api.py
@@ -39,7 +39,7 @@ class TokenRefreshAPI(basehandlers.APIHandler):
   # Note: we use only POST instead of GET to avoid attacks that use GETs.
   def do_post(self, **kwargs):
     """Refresh the session and return a new XSRF token for the current user."""
-    logging.info("TokenRefreshAPI do_post. About to refresh user session.")
+    # logging.info("TokenRefreshAPI do_post. About to refresh user session.")
     user = self.get_current_user()
     users.refresh_user_session()
     result = {
